Support get common models for specific api version (#700)
#### Config
1. A new option `api-version` will be added to the TCGC context and
should be configed in languages' `tspconfig.yaml` file.
2. If the `api-version` is not set or set to `latest`, then all returns
of TCGC APIs will be the information of latest API version.
3. If the `api-version` is set to a specific version, then all returns
of TCGC APIs will be the information of that API version. If the version
is not existed, then a diagnostic warning will be passed to emitter and
will fallback to latest version.
5. If the `api-version` is set to `all`, then the returns of TCGC APIs
will be the information of all API versions. Next part will describe how
the versioning infomation is included in TCGC common models.

---------

Co-authored-by: iscai-msft <43154838+iscai-msft@users.noreply.github.com>