Fix client version params (#778)
Right now, we have all api version information tied to the `sdkContext`
as a whole. As an example, this means that when we add a client in a
later version, we're not able to have granular handling of api version
for the second client.

In this PR, I've made all api version information a mapping from
namespace to apiVersion information. I also added a test that reproduces
one of the previous issues.

The PR diff is large bc I've had to add parameters everywhere so I'm
still passing

---------

Co-authored-by: iscai-msft <isabellavcai@gmail.com>