[tcgc] don't require core-introduced params with `@override` (#1435)
With `@override`, we require a 1:1 mapping between the actual parameters
and the parameters passed in your method signature override. However,
this causes issues with `core`-introduced parameters because of
template-related issues. On a higher level though, I think it's fair to
not require `core` params in the method signature, since we already hide
these parameters in the method signature, and these parameters will
still be eventually set to the service

cc @trangevi

---------

Co-authored-by: iscai-msft <isabellavcai@gmail.com>