[TCGC] Change the `responses` and `exceptions` to be arrays (#1540)
Fixes https://github.com/Azure/typespec-azure/issues/1453

This is a breaking change in TCGC, but there are some reasons of making
this change:
1. The `HttpOperation.responses` from typespec core library also defines
as an array.
2. The type `Map` is not persistable in typescript, which causes issues
in some emitter implementation such as dotnet, because it needs to
serialize everything in TCGC output and sends it to the generator
written in other language (details
[here](https://github.com/Azure/autorest.csharp/issues/5028))
3. In most cases, we have no more than 4 responses corresponding
different status codes. In the majority of cases, we have 2 or 1. In
such a small collections, maps do not have distinguishable performance
improvement comparing with arrays with `find`, and arrays have the
advantages above.