[@azure-tools/azure-http-specs] - Azure Specific Spec Scenarios to test Typespec (#1665)
This is a spin-off of PR https://github.com/Azure/cadl-ranch/pull/741. A
new package `@azure-tools/azure-http-specs` has been created to handle
only the azure specific scenarios and the scenarios that are not
included in the `@typespec/http-specs` package.
Please review and approve the PR. 

Note:
1. The `test:e2e` script is working fine.
2. I have not deleted the `cadl-ranch-specs` folder under `e2e-tests`
folder and not modified the `e2e-tests.mjs` script.