Versioning documentation: Reordering documentation to make tspconfig.yaml change more visible  (#1756)
Working on this issue:
https://github.com/Azure/typespec-azure/issues/1397

The changes are done on `tspconfig.yaml` but is not the first visible
tab so is difficult to observe where the changes are done, so all 3
scenarios look the same, reordering this to show on the first tab the
file in where the changes are done.
 
Creating a draft as I need to corroborate if more changes are needed.