Introduce more usages to describe things related with the states of LRO (#1887)
Remove this workaround since #311 is closed.
Fixes https://github.com/Azure/typespec-azure/issues/1752
Fixes https://github.com/Azure/typespec-azure/issues/1530

This PR adds three new usage flags: `LroInitial`, `LroPolling`,
`LroFinalEnvelope`.
These new usage flags are previously represented by `Output`, now the
`Output` flag no longer contain the semantics of these usages.
Therefore, if a model's usage is only `LroInitial`, this means it is
only used as the initial response of an LRO.
if a model has both `LroInitial` and `Output`, this means it is used as
the initial response of an LRO, and it also used as a normal output in
other places.
The final result will not have above new usages unless it participates
as those roles as well. It will have output usage.