`typespec-azure` repository - Upload Manifest file to the new storage account (#1976)
In the `cadl-ranch` repository, during the publish process, the CI
publishes the manifest file to the `azuresdkcadlranch` storage account.
You can find the code
[here](https://github.com/Azure/cadl-ranch/blob/main/.azure-pipelines/publish.yml#L35)

With the new Azure Spec Dashboard, the `typespec-azure` repository CI
has to publish the manifest file to the `typespec` storage account. I
have tested this code in CI in my branch. You can find the link
[here](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4403105&view=logs&j=eab7c258-2ac5-5cd8-1109-0a296aef509c&t=99d957b2-3f06-5bc3-a8e6-5126b230307b).

Please review and approve the PR. Thanks