Add flavor in Azure tutorial (#2003)
While we have issues to remove `flavor` for `azure-tools` emitters,
matter of fact is that today it's needed, and people got confused
following the tutorial today (for good reasons).
I'll do a PR once we change the default, it may be in several months
given this is not a priority.