Add `@scope` decorator (#1998)
Resolves #964

The scope parameter of `@scope` decorator is not working as the normal
decorator.
For a normal decorator, we need to get the decorator value from state
map based on the scope and the current emitter name. If the value hit
from the state map with current emitter name, it's a positive case, if
not hit, it's a negative case.
```
@clientName("Renamed", csharp)
model A;
```
With the normal decorator scope handling, if the decorator is absent it
is a negative case. For the case below, we don't need to handle
`@clientname` decorator, it is a negative case.
```
model A;
```


But for `@scope` decorator, when it is absent, it means `AllScopes`,
which is also a positive case. For below two cases, python emitter
should include this operation, both of them are positive case.
```
op func: void
```
and 
```
@scope("!csharp")
op func: void
```

For incremental applying, the mechanism should be similar to the normal
decorator scope handling.
- for the same scope, the later decorator value wins regardless it's
defined with normal scope or scope negation