[typespec-autorest] Fix logic around `canShare` and `@invisible` (#2307)
This PR fixes a bug with the handling of `@invisible(Lifecycle)` in
typespec-autorest. Currently, it considers a property shareable if it
either has default visibility (all modifiers), or if it does not contain
any modifiers that cannot be expressed through x-ms-mutability. This was
almost correct, except for the case where the property is _completely
invisible_, in which case the property cannot be shared. This behavior
was unique to typespec-autorest due to its logic around handling
x-ms-mutability and the openapi3 emitter was not affected because it
only attempts to share read-only properties.

In practice, `@invisible(Lifecycle)` is extremely rarely used because it
means that the property would not be present in any operation payloads.

This PR adds a test against regressions ensuring that
`@invisible(Lifecycle)` correctly hides properties.