Change OperationStatusResult.resourceId to armResourceIdentifier (#2435)
Change `OperationStatusResult.resourceId` to `armResourceIdentifier` as
what we did for `OperationStatusResult.id`.

Justification:
Though semantically it is the same as before, in the client we have the
logic "if the string is armResourceIdentifier from Azure.Core, then we
use the ResourceId type from our core library". If
`OperationStatusResult.resourceId` is giving us a `string` type, we
don't know it is actually from `Azure.Core`, so we won't do the
replacement to the core type.

---------

Co-authored-by: Pan Shao <pashao@microsoft.com>
Co-authored-by: Alitzel Mendez <6895254+AlitzelMendez@users.noreply.github.com>