azure-http-specs, fix dyn in handler (#2609)
@timotheeguerin

I've got 500 on this API, in nightly build (against unreleased
azure-http-specs lib -- released one is good)

https://dev.azure.com/azure-sdk/public/_build/results?buildId=4814852&view=logs&j=ca395085-040a-526b-2ce8-bdc85f692774&t=1fb2445c-3c92-5f9b-bcad-710222d02bc3&l=274

So far, this appears to be the only place `dyn/dynItem` is used in
handler. My guess is that this be a mistake.

Let me know if this is the not the correct fix.