[TCGC] Fix missing info for multipart cases after model property refactor (#3012)
This is a behavior breaking change. Currently, TCGC always replaces the
property type of a multipart model with the real body type for the HTTP
part. So, emitter could rely on the property type directly to
serialize/deserialize an HTTP part. Also, the header parameters in one
HTTP part are added to the `headers` property of `MultipartOptions`
type.

For example:
```tsp
model MultiPartRequest {
  stringWithContentType: HttpPart<{
    @body body: string;
    @header contentType: "text/html";
  }>;
}
@post
op upload(
  @header contentType: "multipart/form-data",
  @multipartBody body: MultiPartRequest,
): void;
```
The type of `stringWithContentType` is `string` instead of an anonymous
model `{@body body: string; @header contentType: "text/html";}` as
before.

## Migration Guide
For any property with multipart serialization option, use the property
type directly instead of looking for body parameter as before.