[tcgc] add `@Azure.ClientGenerator.Core.Legacy.legacyHierarchyBuilding` decorator (#3031)
fixes #3015

Some things to note:
1. This is the solution that it appears the TypeSpec team is more in
favor of, bc it doesn't have bearing on the rest api definition
2. For your emitter to respect this decorator, you have to pass in
`respectLegacyHierarchyBuilding` when you call `createSdkContext`
3. I didn't do any automatic property inheritance from parents, I don't
think TCGC should do that. Instead, I do have warnings ensuring that the
models are defined in the raw TypeSpec such that the properties are
correctly inherited, otherwise the `@legacyHiearchyBuilding` decorator
gets ignored.
4. I do some post processing on `.discriminatedSubtypes` on the
`SdkModelType`. I make sure that models from `@ legacyHiearchyBuilding `
are added here. This is for ease of emitters implementing

---------

Co-authored-by: iscai-msft <isabellavcai@gmail.com>
Co-authored-by: Chenjie Shi <tadelesh.shi@live.cn>
Co-authored-by: tadelesh <chenjieshi@microsoft.com>