[autorest] XML emit strategy (#3360)
This PR adds an option for autorest XML emit that allows the project
configuration to disable XML serialization annotations. This can be
useful for some specifications that have operations that return raw XML
payloads without polluting the whole spec with XML annotations. It also
provides a point of extension in the future if more XML strategies are
required.