Adding private experimental decorators. (#3449)
Adding these decorators as private to enable the model extension and
customization required by Foundry. These are the same decorators
described in [this
document](https://github.com/microsoft/openai-openapi-pr/blob/main/docs/foundry-interim.md),
and have been used for development of the specifications in a private
repository.
Now that we are approaching the point where the APIs will go public, we
need to move their development to the public repository. This requires
making these decorators available in a more accessible package.

---------

Co-authored-by: Will Temple <witemple@microsoft.com>