Fix /azure/resource-manager/operation-templates createOrReplace LRO responses (#3517)
Per
https://github.com/Azure/typespec-azure/blob/main/packages/azure-http-specs/specs/azure/resource-manager/operation-templates/lro.tsp#L121-L138,
the final object should blend into the status monitor, and not come as
one of its properties. `method.lroMetadata.finalResultPath` is not
`'properties'` - it is `undefined`, meaning that TCGC tells us it should
come as spread and not enveloped inside the `'properties'` property.

Also fix the mockapi to return HTTP 201 for the first response, and
return 200 for subsequent responses, as per what the docs say -
https://github.com/Azure/typespec-azure/blob/main/packages/azure-http-specs/specs/azure/resource-manager/operation-templates/lro.tsp#L60-L120